Refactor: #123 SettingService의 logout 메서드를 AuthService로 이관 - SettingService 내의 logout()을 AuthService로 이동 - UserRepositoryImpl에서 logout 호출 시 AuthService 사용하도록 수정 - SettingService는 member 패키지로 이동
