+++ app/src/main/java/com/konkuk/medicarecall/data/api/auth/AuthService.kt
@@ -7,6 +7,7 @@ import com.konkuk.medicarecall.data.dto.response.MemberTokenResponseDto
 import com.konkuk.medicarecall.data.dto.response.VerificationResponseDto
 import retrofit2.Response
 import retrofit2.http.Body
+import retrofit2.http.Header
 import retrofit2.http.POST
 
 interface AuthService {
@@ -21,4 +22,9 @@ interface AuthService {
     @POST("verifications/confirmation")
     suspend fun confirmPhoneNumber(@Body req: PhoneNumberConfirmRequestDto)
         : Response<VerificationResponseDto>
+
+    @POST("auth/logout")
+    suspend fun logout(
+        @Header("Authorization") authorization: String // "Bearer <refresh>"
+    ): Response<Unit>
 }
